refactor：types：reducers：types移動
